Move utility code into separate file
The utility code that handles the system preparation tasks and the
"not-so-interesting" things is now contained in the utils.cpp file.
